カメラの回転と移動をわかりにくい操作方法で実装

+++ app/src/main/cpp/renderer.h
@@ -12,7 +12,7 @@ void on_update();
 void on_draw();
 void on_pause();
 void on_resume();
-void move_camera(int x, int y);
-void rotate_camera(int rotation);
+void move_camera(float x, float y);
+void rotate_camera(float rotation);
 
 #endif //OPENGL_STUDY_RENDERER_H
